move initial seeding ffunction to simulator class
