+++ experiments/functions.py
@@ -34,14 +34,14 @@ def _run_opencl_model(i, iterations, snapshot_filepath, params, opencl_dir, num_
     # set the random seed of the model for each repetition, otherwise it is completely deterministic
     snapshot.seed_prngs(i)
 
-    # seed initial infections using GAM initial cases
-    data_dir = os.path.join(opencl_dir, "data")
-    snapshot.seed_initial_infections(num_seed_days=num_seed_days, data_dir=data_dir)
-
     # Create a simulator and upload the snapshot data to the OpenCL device
     kernel_dir = os.path.join(opencl_dir, "ramp", "kernels")
     simulator = Simulator(snapshot, kernel_dir=kernel_dir, gpu=use_gpu)
     simulator.upload_all(snapshot.buffers)
+
+    # seed initial infections using GAM initial cases
+    data_dir = os.path.join(opencl_dir, "data")
+    simulator.seed_initial_infections(num_seed_days=num_seed_days, data_dir=data_dir)
     
     print(f"Running simulation {i+1}.")
     summary, final_state = run_headless(simulator, snapshot, iterations, quiet=True,
