+++ microsim/main.py
@@ -233,17 +233,13 @@ def run_opencl_model(individuals_df, activity_locations, time_activity_multiplie
         if disease_params["improve_health"]:
             print("Switching to healthier population")
             snapshot.switch_to_healthier_population()
-
-    # seed initial infections using GAM initial cases
-    snapshot.seed_initial_infections(num_seed_days=disease_params["seed_days"])
-
     if initialise:
         print("Have finished initialising model. -init flag is set so not running it. Exitting")
         return
 
     run_mode = "GUI" if use_gui else "headless"
     print(f"\nRunning OpenCL model in {run_mode} mode")
-    run_opencl(snapshot, iterations, data_dir, use_gui, use_gpu, quiet=False)
+    run_opencl(snapshot, iterations, data_dir, use_gui, use_gpu, disease_params["seed_days"], quiet=False)
 
 
 def run_python_model(individuals_df, activity_locations_df, time_activity_multiplier, msim_args, iterations,
