+++ microsim/opencl/ramp/run.py
@@ -10,7 +10,7 @@ from microsim.opencl.ramp.summary import Summary
 from microsim.opencl.ramp.disease_statuses import DiseaseStatus
 
 
-def run_opencl(snapshot, iterations=100, data_dir="./data", use_gui=True, use_gpu=False, quiet=False):
+def run_opencl(snapshot, iterations=100, data_dir="./data", use_gui=True, use_gpu=False, num_seed_days=5, quiet=False):
     """
     Entry point for running the OpenCL simulation either with the UI or in headless mode.
     NB: in order to write output data for the OpenCL dashboard you must run in headless mode.
@@ -22,6 +22,7 @@ def run_opencl(snapshot, iterations=100, data_dir="./data", use_gui=True, use_gp
     # Create a simulator and upload the snapshot data to the OpenCL device
     simulator = Simulator(snapshot, use_gpu)
     simulator.upload_all(snapshot.buffers)
+    simulator.seed_initial_infections(num_seed_days=num_seed_days)
     if not quiet:
         print(f"Platform:\t{simulator.platform_name()}\nDevice:\t\t{simulator.device_name()}\n")
