+++ microsim/opencl/ramp/snapshot.py
@@ -152,39 +152,6 @@ class Snapshot:
 
         return cls(nplaces, npeople, nslots, time, area_codes, not_home_probs, lockdown_multipliers, buffers)
 
-    def seed_initial_infections(self, num_seed_days=5, data_dir="microsim/opencl/data/"):
-        """
-        Seeds initial infections by assigning initial cases based on the GAM assigned cases data.
-        The cases for the first num_seed_days days are all seeded at once, eg. they are in the snapshot before the
-        simulation is run.
-        Initial cases are assigned to people from higher risk area codes who spend more time outside of their home.
-        """
-
-        # load initial case data
-        initial_cases = pd.read_csv(os.path.join(data_dir, "devon_initial_cases.csv"))
-        num_cases = initial_cases.loc[:num_seed_days - 1, "num_cases"].sum()
-
-        msoa_risks_df = pd.read_csv(os.path.join(data_dir, "msoas.csv"), usecols=[1, 2])
-
-        # combine into a single dataframe to allow easy filtering based on high risk area codes and
-        # not home probabilities
-        people_df = pd.DataFrame({"area_code": self.area_codes, "not_home_prob": self.not_home_probs})
-        people_df = people_df.merge(msoa_risks_df, on="area_code")
-
-        # get people_ids for people in high risk MSOAs and high not home probability
-        high_risk_ids = np.where((people_df["risk"] == "High") & (people_df["not_home_prob"] > 0.3))[0]
-
-        # randomly choose a given number of cases from the high risk people ids.
-        initial_case_ids = np.random.choice(high_risk_ids, num_cases)
-
-        # Assign initial cases by updating people_statuses buffer
-        self.buffers.people_statuses[initial_case_ids] = DiseaseStatus.Exposed.value
-
-        # Set initial transition times to 1
-        self.buffers.people_transition_times[initial_case_ids] = 1
-
-        self.time = num_seed_days
-
     def update_params(self, new_params):
         try:
             self.buffers.params[:] = new_params.asarray()
